Add objectId verifications

+++ api/src/service/cities-service.ts
@@ -1,5 +1,7 @@
 import { CitiesRepository } from '../repository/cities-repository';
 import { ICity } from '../models/City';
+import { isObjectId } from '../core/Utils';
+import { BadRequestError } from '../core/CustomError';
 
 export class CitiesService {
     static async createCity(cityData: Partial<ICity>): Promise<ICity> {
@@ -7,6 +9,9 @@ export class CitiesService {
     }
 
     static async getCityById(cityId: string): Promise<ICity | null> {
+        if (!isObjectId(cityId)) {
+            throw new BadRequestError('cityId is not a valid ObjectId');
+        }
         return await CitiesRepository.readById(cityId);
     }
 
@@ -15,10 +20,16 @@ export class CitiesService {
     }
 
     static async updateCity(cityId: string, cityData: Partial<ICity>): Promise<ICity | null> {
+        if (!isObjectId(cityId)) {
+            throw new BadRequestError('cityId is not a valid ObjectId');
+        }
         return await CitiesRepository.update(cityId, cityData);
     }
 
     static async deleteCity(cityId: string): Promise<ICity | null> {
+        if (!isObjectId(cityId)) {
+            throw new BadRequestError('cityId is not a valid ObjectId');
+        }
         return await CitiesRepository.delete(cityId);
     }
 }
