+++ api/src/service/locations-service.ts
@@ -1,5 +1,7 @@
 import { ILocation } from '../models/Location';
 import { LocationsRepository } from '../repository/locations-repository';
+import { isObjectId } from '../core/Utils';
+import { BadRequestError } from '../core/CustomError';
 
 export class LocationsService {
     static async createLocation(locationData: Partial<ILocation>): Promise<ILocation> {
@@ -7,6 +9,9 @@ export class LocationsService {
     }
 
     static async getLocationById(locationId: string): Promise<ILocation | null> {
+        if (!isObjectId(locationId)) {
+            throw new BadRequestError('locationId is not a valid ObjectId');
+        }
         return await LocationsRepository.readById(locationId);
     }
 
@@ -15,10 +20,16 @@ export class LocationsService {
     }
 
     static async updateLocation(locationId: string, locationData: Partial<ILocation>): Promise<ILocation | null> {
+        if (!isObjectId(locationId)) {
+            throw new BadRequestError('locationId is not a valid ObjectId');
+        }
         return await LocationsRepository.update(locationId, locationData);
     }
 
     static async deleteLocation(locationId: string): Promise<ILocation | null> {
+        if (!isObjectId(locationId)) {
+            throw new BadRequestError('locationId is not a valid ObjectId');
+        }
         return await LocationsRepository.delete(locationId);
     }
 }
